Cleanup: Remove wss://*.stream.<DOMAIN>:4443 from connect_src
From the testing done, it seems allowing wss://*.stream.<DOMAIN>:4443 in connect_src does nothing because Hubs isn't currently using any subdomains below stream, plus it theoretically shouldn't need to be specified here because it is a URL local to the cluster.
So far no issues have been observed with removing it.