Add a maintenance mode to the restore backup script
What: Adds a primitive maintenance mode to the restore backup script.  This is applied at the beginning and sets haproxy to redirect traffic to a non-existent maintenance mode subdomain, then restarts the instance to totally disconnect any people present on the instance and prevent anyone new from joining.  The redirects are removed and the instance is returned to normal at the end after the restore is finished.

Why: So that people can't interrupt/corrupt the restore by modifying data on the instance while the restore is happening.

Note: At present the maintenance mode isn't a real page, so it's not all that pretty, and you won't be redirected back to your previous page once the restore is finished (even if you reload), but it gets the job done.  Ideally, these faults should be addressed at some point in the future.