Keep the user apprised of the progress of the restore backup script
What: Prints headings for the general steps of the restore script and prints the command output to the terminal.

Why: This is a very involved and potentially long running script, and the additional output should help reduce confusion as to whether the script is running normally or has gotten stuck.

Note: This implements similar behavior to the apply script, but that, at present, will only work with the main configuration file and not a secondary, temporary one.  In the future, the code to apply a Kubernetes configuration and monitor the deployment status should potentially be further abstracted so it can work with any configuration files and only one version of the code is needed.