Change Community Edition license to MPL-2.0
What: Changes the license for the Community Edition Node scripts to be MPL-2.0, the same as the top level license on this repository.

Why: The license should always have been MPL-2.0, but it was missed that the license wasn't MPL-2.0 when the initial pull request that added the Node scripts was reviewed.