Fix kubectl cp sometimes failing (and not retrying) in version 1.30+
What: Passes an environment variable to the kubectl cp command to disable using websockets.

Why: Websockets are enabled by default in kubectl 1.30+ and this can cause transfers to fail and not retry.  Disabling websockets avoids the issue.

Referrences:
Link to GitHub issue with the documented workaround:
https://github.com/kubernetes/kubernetes/issues/60140#issuecomment-2565537460

Link to GitHub PR which introduced websockets as the default and the note that it affects kubectl cp:
https://github.com/kubernetes/kubernetes/pull/123281#issuecomment-2647550899