Fix PostgreSQL data persistence
What:
Returns the mountPath for the postgresql-data volume back to it's original value and adds the PGDATA environment variable to the container to specify that the PostgreSQL data be stored in a further subfolder.

Why:
Changing the mountPath doesn't actually change where the data is written, instead it causes Docker to create an anonymous volume at the default path that is discarded when the container/pod is deleted (which loses all the data).

Using the PGDATA environment variable to specify that the PostgreSQL data be stored in a further subfolder should continue to avoid the issue of a lost+found folder being created somewhere PostgreSQL doesn't like it.

Note:
This will need to be updated in the future because future versions of PostgreSQL expect the data to be stored in version specific folders.